backend/requirements.txt
# Core dependencies
apache-airflow==2.10.4
fastapi==0.115.5
httpx==0.28.1
python-dotenv==1.2.1
uvicorn[standard]==0.32.1
email-validator>=2.0.0
pydantic-settings==2.6.1